trim InfoPage to only contain Cart functionality
