+++ frontend/src/index.js
@@ -4,7 +4,14 @@ import 'react-app-polyfill/stable'
 
 import ReactDOM from 'react-dom'
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom'
-import { AboutPage, AuthPage, ErrorPage, FAQPage, InfoPage, ReviewPage } from './pages'
+import {
+  AboutPage,
+  AuthPage,
+  CartPage,
+  ErrorPage,
+  FAQPage,
+  ReviewPage,
+} from './pages'
 import { GoogleAnalytics } from './components/common'
 
 if (window.location.hostname !== 'localhost') {
@@ -19,11 +26,11 @@ ReactDOM.render(
       <Route exact path="/" component={ReviewPage} />
       <Route path="/about" component={AboutPage} />
       <Route path="/faq" component={FAQPage} />
+      <Route path="/cart" component={CartPage} />
       <Route
         path="/:type(course|department|instructor)/:code"
         component={AuthPage}
       />
-      <Route path="/:page(cart)" component={InfoPage} />
       <Route component={ErrorPage} />
     </Switch>
     <GoogleAnalytics />
