+++ frontend/src/pages/index.js
@@ -2,6 +2,6 @@ export * from './AboutPage'
 export * from './AuthPage'
 export * from './ErrorPage'
 export * from './FAQPage'
-export * from './InfoPage'
+export * from './CartPage'
 export * from './ReviewPage'
 export * from './Tokenform'
